+++ src/Downloader/ChunkDownloader.cs
@@ -21,7 +21,7 @@ namespace Downloader
         protected int BufferBlockSize { get; }
         public event EventHandler<DownloadProgressChangedEventArgs> DownloadProgressChanged;
 
-        public async Task<Chunk> Download(Request downloadRequest, int maximumSpeed, CancellationToken token)
+        public async Task<Chunk> Download(Request downloadRequest, long maximumSpeed, CancellationToken token)
         {
             try
             {
@@ -53,7 +53,7 @@ namespace Downloader
             }
         }
 
-        private async Task DownloadChunk(Request downloadRequest, int maximumSpeed, CancellationToken token)
+        private async Task DownloadChunk(Request downloadRequest, long maximumSpeed, CancellationToken token)
         {
             if (token.IsCancellationRequested ||
                 IsDownloadCompleted())
